Fix selector recalculating each time
The problem was that a new object was being created each time in the
initial selectors. This still doesn't solve the problem of every time
*any* entity updates the spatial array, *all* the selectors in
`world.js` are renegerated (and therefore *all* the `useEffect` handlers
are called too).

We might need another layer of selectors which calculate whether
something has actually changed which means we need a redraw, or perhaps
we keep a manual store of the renderables and manage diffs ourselves.