Filter actions with habitable requirements
I'm pretty unhappy with the habitable filtering code in `world.js`. At
the least, it needs tests. It's hard to understand and therefore prone
to bugs. There are definitely better ways of checking which entities
exist on the same tile than relying on a side effect of the action
compilation.