Move actions to state + intro rules engine
createSelector is only reeally meant to be used for transforming
existing state, not layer on new state in an async fashion (like running
a rules engine). Therefore we do that in the component itself using
useState and useEffect.

We're using json-rules-engine-simplified for checking "can we run this
action" and "what does this action do when chosen". Seems to be working
ok so far.

This is now a much improved setup to previous filtering checks for actions.