Remove redux
Decided to bite the bullet and go for more of an imperative shell.
Reasons were mostly that trying to build a Redux app with an ECS meant
an awful lot of custom selectors. I'd also have had to manage the pixi
state through diffing the ECS and applying changes: I'd prefer to work
directly with linked display objects.

It isn't quite as nice: the dispatching of actions from the UI is
currently fairly nasty, and as I have to store the state itself in a ref
object, I need to force a UI re-render whenever I want a new UI state. I
could move the UI state to a more declarative object, but then the UI
wouldn't be able to query the ECS directly. This feels like an ok
compromise.