Explore the map
We've introduced the ability for the character to see other locations
outside the starting hex which is introducing a bunch of new problems,
such as the allowed distance for each action.

Also our `generateActions` code assumes the player, rather than the
specific person we click on.  Once we start seeing other characters,
we're going to have to redo that UI code.