BIG HTN planner upgrade - eat now works
Improved with expected effects in the primitives, and allowed a layer of
indirection when walking somewhere. There are now breaks in the finding
of places: if we cannot find a place to do something, we don't try again
for an hour and fall down to the next high level task.

Closes #37.