Move selector to async state call in map
We cannot call an async function like Engine.run(...) inside the selector. This is problematic, as either we have to move this to a redux thunk using createAsyncThunk(...), or we move it to the component and calculate each time the tiles change. This works, as it's the equivalent to the derived createSelector, with the problem that it's only available to the <Map/> component (which we can live with for now).