Refactor player choice to terrain not tile
TODO: choose an event / tile to actually place here, rather than just
the terrain. The idea is that giving the player some agency to choose
the terrain means they get to shape the land they want, BUT they don't
get to choose the event, which will come from the various conditions on
the tiles themselves.

Got to this point by realising that arbitrary restrictions on placement
were hurting the fun - it's just a frustrating puzzle then rather than
something creative.